:truck: [#1546] -- Move logic rule order tests into test module for deterministic evaluation
It's not really related to variables per se, the variables are just
a vehicle to make assertions about the order of evaluation.
