+++ src/openforms/submissions/tests/form_logic/test_modify_variables.py
@@ -246,53 +246,3 @@ class VariableModificationTests(VariablesTestMixin, TestCase):
         evaluate_form_logic(submission, submission_step2, submission.data)
 
         self.assertEqual(7, submission_step2.data["nTotalBoxes"])
-
-    def test_logic_rule_order_respected(self):
-        form = FormFactory.create()
-        step = FormStepFactory.create(
-            form=form, form_definition__configuration={"components": []}
-        )
-        FormVariableFactory.create(
-            form=form,
-            key="a",
-            source=FormVariableSources.user_defined,
-            data_type=FormVariableDataTypes.int,
-        )
-        FormLogicFactory.create(
-            form=form,
-            order=1,
-            json_logic_trigger={"==": [1, 1]},
-            actions=[
-                {
-                    "variable": "a",
-                    "action": {
-                        "name": "Add 2",
-                        "type": "variable",
-                        "value": {"+": [{"var": "a"}, 2]},
-                    },
-                }
-            ],
-        )
-        FormLogicFactory.create(
-            form=form,
-            order=0,
-            json_logic_trigger={"==": [1, 1]},
-            actions=[
-                {
-                    "variable": "a",
-                    "action": {
-                        "name": "Multiply by 2",
-                        "type": "variable",
-                        "value": {"*": [{"var": "a"}, 2]},
-                    },
-                }
-            ],
-        )
-        submission = SubmissionFactory.create(form=form)
-        submission_step = SubmissionStepFactory.create(
-            submission=submission, form_step=step, data={"a": 3}
-        )
-
-        evaluate_form_logic(submission, submission_step, {"a": 3})
-
-        self.assertEqual(submission_step.data["a"], 8)  # ( 3 x 2 ) + 2
